refactor: move event models into models.events with shims
